migrate ir and all backends to transport-based codegen

+++ boltffi_bindgen/src/ir/codec.rs
@@ -115,7 +115,7 @@ pub enum EnumLayout {
 #[derive(Debug, Clone)]
 pub struct VariantLayout {
     pub name: VariantName,
-    pub discriminant: i64,
+    pub discriminant: i128,
     pub payload: VariantPayloadLayout,
 }
